requirements.txt
Jinja2==3.0.3
MarkupSafe==2.1.0
PyYAML==6.0
